Rename pkg for resourceClusterHost classes
