+++ mantis-control-plane/mantis-control-plane-server/src/main/java/io/mantisrx/master/resourcecluster/proto/MantisResourceClusterEnvType.java
@@ -14,15 +14,11 @@
  * limitations under the License.
  */
 
-package io.mantisrx.control.plane.resource.cluster.resourceprovider;
+package io.mantisrx.master.resourcecluster.proto;
 
-import io.mantisrx.control.plane.resource.cluster.proto.ResourceClusterProvisionSubmissionResponse;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class NoopResourceClusterResponseHandler implements ResourceClusterResponseHandler {
-    @Override
-    public void handleProvisionResponse(ResourceClusterProvisionSubmissionResponse resp) {
-        log.info(resp.toString());
-    }
+public enum MantisResourceClusterEnvType {
+    Default,
+    Dev,
+    Test,
+    Prod
 }
