+++ mantis-control-plane/mantis-control-plane-server/src/main/java/io/mantisrx/master/resourcecluster/proto/ResourceClusterProvisionSubmissionResponse.java
@@ -14,11 +14,18 @@
  * limitations under the License.
  */
 
-package io.mantisrx.control.plane.resource.cluster.proto;
+package io.mantisrx.master.resourcecluster.proto;
 
-public enum MantisResourceClusterEnvType {
-    Default,
-    Dev,
-    Test,
-    Prod
+import lombok.Builder;
+import lombok.Value;
+
+/**
+ * Response contract for {@IResourceClusterProvider} provision request.
+ */
+@Builder
+@Value
+public class ResourceClusterProvisionSubmissionResponse {
+    String response;
+
+    Throwable error;
 }
