+++ mantis-control-plane/mantis-control-plane-server/src/main/java/io/mantisrx/master/resourcecluster/resourceprovider/ResourceClusterResponseHandler.java
@@ -14,27 +14,13 @@
  * limitations under the License.
  */
 
-package io.mantisrx.control.plane.resource.cluster.proto;
+package io.mantisrx.master.resourcecluster.resourceprovider;
 
-import io.mantisrx.shaded.com.google.common.base.Joiner;
-import lombok.Builder;
-import lombok.Value;
+import io.mantisrx.master.resourcecluster.proto.ResourceClusterProvisionSubmissionResponse;
 
-@Builder
-@Value
-public class ScaleResourceRequest {
-    String clusterId;
-
-    String skuId;
-
-    String region;
-
-    MantisResourceClusterEnvType envType;
-
-    int desireSize;
-
-    public String getScaleRequestId() {
-        return Joiner.on('-').join(this.clusterId, this.region, this.envType.name(), this.skuId, this.desireSize);
-    }
-
+/**
+* Callback handler for {@link ResourceClusterProvider} responses.
+*/
+public interface ResourceClusterResponseHandler {
+    void handleProvisionResponse(ResourceClusterProvisionSubmissionResponse resp);
 }
